[LLT-5564] Move MQTT listener to the test
